4264 Extend TransactionLogSnapshotMetadataTest
