--- java/statestorev2/src/test/java/sleeper/statestorev2/transactionlog/snapshots/TransactionLogSnapshotMetadataBasePathTest.java
@@ -1,46 +1,0 @@
-/*
- * Copyright 2022-2025 Crown Copyright
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package sleeper.statestorev2.transactionlog.snapshots;
-
-import org.junit.jupiter.api.Test;
-
-import sleeper.core.properties.instance.InstanceProperties;
-import sleeper.core.properties.table.TableProperties;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.core.properties.instance.CdkDefinedInstanceProperty.DATA_BUCKET;
-import static sleeper.core.properties.instance.CommonProperty.FILE_SYSTEM;
-import static sleeper.core.properties.table.TableProperty.TABLE_ID;
-import static sleeper.core.properties.testutils.InstancePropertiesTestHelper.createTestInstanceProperties;
-import static sleeper.core.properties.testutils.TablePropertiesTestHelper.createTestTableProperties;
-import static sleeper.core.schema.SchemaTestHelper.createSchemaWithKey;
-
-public class TransactionLogSnapshotMetadataBasePathTest {
-
-    @Test
-    void shouldCreateBasePathForTable() {
-        // Given
-        InstanceProperties instanceProperties = createTestInstanceProperties();
-        instanceProperties.set(FILE_SYSTEM, "file://");
-        instanceProperties.set(DATA_BUCKET, "test-bucket");
-        TableProperties tableProperties = createTestTableProperties(instanceProperties, createSchemaWithKey("key"));
-        tableProperties.set(TABLE_ID, "test-table-id");
-
-        // When / Then
-        assertThat(TransactionLogSnapshotMetadata.getBasePath(instanceProperties, tableProperties))
-                .isEqualTo("file://test-bucket/test-table-id");
-    }
-}
